+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### :sparkles: Release Note <!-- optional -->
 
 ### :rocket: Enhancements
+
+- [#842](https://github.com/reviewdog/reviewdog/pull/842) Incude suggestion as markdown diff while making Gitlab merge request comment
 - ...
 
 ### :bug: Fixes